factored out randrange to its own utils file
