+++ src/lib/mapGeneration/mapUtils.h
@@ -1,0 +1,3 @@
+#pragma once
+
+int randRange(int min, int max);
